refactor(lists): migrate hierarchyView.js to TypeScript
Мигрировал hierarchyView.js (816 строк) в модульную TypeScript структуру listsManager/rendering/.

Созданные файлы:
- types/hierarchy.d.ts: Типы для иерархии (HierarchyTree, HierarchyStore, HierarchyProductGroup, HierarchyItem, ListsManagerProxy)
- rendering/SwipeHandler.ts: Обработка swipe-жестов с полной типизацией (320 строк)
- rendering/HierarchyView.ts: Рендеринг дерева иерархии Store → ProductGroup → Items (480 строк)
- adapters/hierarchyExports.ts: Экспорт HierarchyView в window для backward compatibility

Обновлённые файлы:
- rendering/hierarchyIntegration.ts: Использование TypeScript классов вместо window.HierarchyView
- index.ts: Экспорт новых классов и типов
- lists-bundle.ts: Импорт hierarchyExports вместо legacy JS файла

Удалённые файлы:
- lists/hierarchyView.js (legacy)
- lists/hierarchyView.min.js
- lists/hierarchyView.min.js.map

Критические детали сохранены:
- Movement detection >5px для отличия tap от swipe
- Transform removal через removeProperty() вместо translateX(0)
- Click blocking в правой зоне (120px) для swipe indicator
- Modal tracking через modalOpenedBySwipe
- XSS защита через escapeHtml() для всех user inputs
- Форматирование количества (кг: 1 decimal, остальные: integer)

Валидация:
- npm run type-check: 0 ошибок
- npm run build: успешно (210.64 kB lists.min.js)

BREAKING: Нет - HTML шаблоны работают без изменений благодаря adapters
